refactor: remove receipt storage adapter mock
