+++ tap_core/src/adapters/mock.rs
@@ -3,4 +3,3 @@
 
 pub mod auditor_executor_mock;
 pub mod executor_mock;
-pub mod receipt_storage_adapter_mock;
